refactor: TryContext를 global 패키지로 이동하고 TryId 생성 로직 통합
- TryContext 이동: core/rest → core/global/tryit/context
- TryIdResolver 추가: UUID 생성 및 Try 요청 검증 로직 통합
- WebSocket → REST 의존성 제거
- 8개 파일 import 경로 업데이트

아키텍처 개선:
- 프로토콜 독립적인 컴포넌트를 global 패키지로 이동
- 중복된 TryId 생성 로직 제거 (REST, WebSocket)
- 명확한 책임 분리: identification vs context management
