+++ backend/src/main/java/kr/co/ouroboros/core/global/tryit/context/package-info.java
@@ -1,0 +1,14 @@
+/**
+ * Try context management using OpenTelemetry Baggage.
+ * <p>
+ * This package provides protocol-agnostic Try context utilities shared by REST and WebSocket.
+ * <p>
+ * <b>Components:</b>
+ * <ul>
+ *   <li>{@link kr.co.ouroboros.core.global.tryit.context.TryContext} - Thread-safe tryId management via OpenTelemetry Baggage</li>
+ * </ul>
+ *
+ * @since 0.0.1
+ * @see kr.co.ouroboros.core.global.tryit.TryHeaders
+ */
+package kr.co.ouroboros.core.global.tryit.context;
