+++ backend/src/main/java/kr/co/ouroboros/core/global/tryit/identification/package-info.java
@@ -1,0 +1,15 @@
+/**
+ * Try request identification utilities.
+ * <p>
+ * This package provides protocol-agnostic utilities for identifying and managing Try requests.
+ * <p>
+ * <b>Components:</b>
+ * <ul>
+ *   <li>{@link kr.co.ouroboros.core.global.tryit.identification.TryIdResolver} - TryId generation and validation</li>
+ * </ul>
+ *
+ * @since 0.0.1
+ * @see kr.co.ouroboros.core.global.tryit.TryHeaders
+ * @see kr.co.ouroboros.core.global.tryit.context.TryContext
+ */
+package kr.co.ouroboros.core.global.tryit.identification;
