+++ backend/src/main/java/kr/co/ouroboros/core/rest/tryit/identification/TryResponseAdvice.java
@@ -1,7 +1,7 @@
 package kr.co.ouroboros.core.rest.tryit.identification;
 
 import kr.co.ouroboros.core.global.tryit.TryHeaders;
-import kr.co.ouroboros.core.rest.tryit.infrastructure.instrumentation.context.TryContext;
+import kr.co.ouroboros.core.global.tryit.context.TryContext;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.core.MethodParameter;
 import org.springframework.core.Ordered;
