--- backend/src/main/java/kr/co/ouroboros/core/rest/tryit/infrastructure/instrumentation/context/package-info.java
@@ -1,15 +1,0 @@
-/**
- * Try context management components.
- * <p>
- * This package contains components for managing Try context using
- * OpenTelemetry Baggage for thread-safe context propagation.
- * <p>
- * <b>Components:</b>
- * <ul>
- *   <li>{@link kr.co.ouroboros.core.rest.tryit.infrastructure.instrumentation.context.TryContext} - Utility for managing tryId in OpenTelemetry Baggage</li>
- * </ul>
- *
- * @since 0.0.1
- */
-package kr.co.ouroboros.core.rest.tryit.infrastructure.instrumentation.context;
-
