+++ backend/src/main/java/kr/co/ouroboros/core/rest/tryit/infrastructure/instrumentation/sampler/TryOnlySampler.java
@@ -7,7 +7,7 @@ import io.opentelemetry.sdk.trace.data.LinkData;
 import io.opentelemetry.sdk.trace.samplers.Sampler;
 import io.opentelemetry.sdk.trace.samplers.SamplingResult;
 import kr.co.ouroboros.core.global.tryit.TryHeaders;
-import kr.co.ouroboros.core.rest.tryit.infrastructure.instrumentation.context.TryContext;
+import kr.co.ouroboros.core.global.tryit.context.TryContext;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.context.request.RequestContextHolder;
 import org.springframework.web.context.request.ServletRequestAttributes;
