+++ backend/src/main/java/kr/co/ouroboros/core/websocket/tryit/identification/TryStompChannelInterceptor.java
@@ -1,7 +1,8 @@
 package kr.co.ouroboros.core.websocket.tryit.identification;
 
 import io.opentelemetry.context.Scope;
-import kr.co.ouroboros.core.rest.tryit.infrastructure.instrumentation.context.TryContext;
+import kr.co.ouroboros.core.global.tryit.context.TryContext;
+import kr.co.ouroboros.core.global.tryit.identification.TryIdResolver;
 import kr.co.ouroboros.core.websocket.tryit.common.TryStompHeaders;
 import kr.co.ouroboros.core.websocket.tryit.infrastructure.messaging.TryPublisherNotifier;
 import lombok.RequiredArgsConstructor;
@@ -43,7 +44,7 @@ public class TryStompChannelInterceptor implements ChannelInterceptor {
 
         // Only inbound channel (client -> server) messages are processed. Additional filtering by command is possible.
         String tryHeader = getFirstNativeHeader(accessor, TryStompHeaders.TRY_HEADER);
-        boolean tryRequested = TryStompHeaders.TRY_HEADER_ENABLED_VALUE.equalsIgnoreCase(tryHeader);
+        boolean tryRequested = TryIdResolver.isTryRequest(tryHeader);
 
         UUID tryId = resolveTryId(accessor, tryRequested);
         if (tryId == null) {
@@ -51,7 +52,7 @@ public class TryStompChannelInterceptor implements ChannelInterceptor {
             return message;
         }
 
-        Scope scope = TryContext.setTryId(tryId);
+        Scope scope = TryIdResolver.activateTryScope(tryId);
         if (scope != null) {
             accessor.setHeader(TryStompHeaders.INTERNAL_SCOPE_HEADER, scope);
         }
@@ -104,7 +105,7 @@ public class TryStompChannelInterceptor implements ChannelInterceptor {
 
         // 2) If Try request header is enabled, generate a new one
         if (tryRequested) {
-            UUID newTryId = UUID.randomUUID();
+            UUID newTryId = TryIdResolver.generateTryId();
             log.debug("Detected STOMP Try request and generated new tryId: {}", newTryId);
             return newTryId;
         }
