+++ backend/src/test/java/kr/co/ouroboros/core/websocket/tryit/identification/TryStompChannelInterceptorTest.java
@@ -1,7 +1,7 @@
 package kr.co.ouroboros.core.websocket.tryit.identification;
 
 import io.opentelemetry.context.Scope;
-import kr.co.ouroboros.core.rest.tryit.infrastructure.instrumentation.context.TryContext;
+import kr.co.ouroboros.core.global.tryit.context.TryContext;
 import kr.co.ouroboros.core.websocket.tryit.common.TryStompHeaders;
 import kr.co.ouroboros.core.websocket.tryit.infrastructure.messaging.TryPublisherNotifier;
 import org.junit.jupiter.api.BeforeEach;
